Only read session-relevant meta config
Starting in v1.9.2, the presence of a `meta` element on any model config
would cause the Glue client to create a new Glue session, even if the
`mdta` element did not contain any parameters relevant to Glue sessions.

The `meta` element on dbt models is generic, and can be used by
different plugins to introduce different behavior. Coupling the mere
presence of `meta` to separate Glue sessions is problematic for users on
v1.9.0 and earlier who were using `meta` for other purposes, because
new sessions are spun up rather than reused, resulting in longer
runtimes, and increased cardinality of sessions IDs propagates to
CloudWatch metric cardinality.

This commit changes the interpretation of `meta` such that new sessions
are only created if `meta` contains config keys that actually affect
Glue session configuration.

Additionally, it appears that the existing unit tests did not cover the
behavior of Glue session creation, making it difficult to verify changes
in behavior for Glue session arguments. This commit sets up spies to
observe the parameters passed to `boto3.GlueClient.create_session` in
unit tests. The previous test for the `meta` parameter has been replaced
because it was testing input configuration and not outgoing behavior.